Funcionalidades Estudiante 3

+++ sgc/src/main/java/co/edu/eam/sistemasdistribuidos/sgc/core/repositories/BorrowRequestRepository.java
@@ -2,8 +2,10 @@ package co.edu.eam.sistemasdistribuidos.sgc.core.repositories;
 
 import co.edu.eam.sistemasdistribuidos.sgc.core.models.BorrowRequest;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface BorrowRequestRepository extends JpaRepository<BorrowRequest, Integer> {
+public interface BorrowRequestRepository extends JpaRepository<BorrowRequest, Long> {
+
 }
